+++ lib/widgets/shared/simpleMesechetWidget.dart
